change to datetime-esque graphs
Monthly downloads and submissions graphs now use unix-timestamps as their x-axis, automating the tick marks generated for it.

Hourly Downloads is exempt due to the simplicity of the data.

A known fault in doing so within a vertical bar graph is the inability to set uniform bar widths, which are partially based on the length of the month -- so the eagle eyed may notice some differing gap space in between certain months.

Monthly downloads has an extra feature to toggle in between bar and line graphs.

If we decide our old method is worth using, it remains as an html page for inspection.

Relevant JIRA ticket: [ARXIVCE-2408]