Parameterized API, and models file.
Rewrote the API to accept parameters in the following format:
/api/get_data?model=XXXX&group_by=YYYY&by_year=BOOL

where
XXXX corresponds to the sheet we want to look at from the database
YYYY chooses column to aggregate our download counts by
BOOL, a optional boolean argument to further aggregate results by year, and appending a year value to the output.

For ease of use and readbility, ORM models now live in models.py, along with a function to return the specified model in it.

Finally landing.html and associated pages has negligible change to address the new API format.

Part of ongoing efforts towards [ARXIVCE-2415].