convert get_todays_downloads to get_daily_downloads
Rewrote the today's downloads backend and frontend code to query a specific date's worth of hourly download statistics.

Refactored the frontend to allow the user to click back/forward a day, as well as select from a calendar for a specific date's info, as well as be more clear in its labelling of hours.